Move sync code to pkg/skaffold/sync/kubectl
This is the correct place for this code. Might require some
reorganization if we add other syncers, like rsync, but that shouldn't
be difficult.
